requirements.txt
# Data Processing
numpy==1.24.3
pandas==2.0.3

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1

# Databases
psycopg2-binary==2.9.9
elasticsearch==8.11.0
neo4j==5.14.1

# Streaming
kafka-python==2.0.2

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Explainability
shap==0.43.0
lime==0.2.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utils
python-dotenv==1.0.0
loguru==0.7.2
requests==2.31.0